Move user agent constant out of _shared
